backend/requirements.txt
fastapi
uvicorn
pydantic
pyserial
sqlalchemy
python-dateutil
